try to fix first hoc and add second hoc
