+++ src/components/main/main.js
@@ -4,6 +4,7 @@ import {connect} from 'react-redux';
 import classNames from 'classnames';
 import ActionCreator from '../../actions/action-creator';
 import withActiveItem from "../../hocs/with-active-item/with-active-item";
+import withMap from "../../hocs/with-map/with-map";
 import PreviewPlaces from '../preview-places/preview-places';
 import Map from '../map/map';
 import Cities from '../cities/cities';
@@ -11,8 +12,9 @@ import PlaceFilter from '../place-filter/place-filter';
 
 const CitiesWrappedHoc = withActiveItem(Cities);
 const PreviewPlacesWrappedHoc = withActiveItem(PreviewPlaces);
+const MapWrapperHoc = withMap(Map);
 
-const Main = ({offers, currentCity, getCities, onGetCity, onGetActiveItem}) => {
+const Main = ({offers, currentCity, getCities, onGetCurrentCity, onGetCurrentOffer}) => {
   const cities = getCities();
 
   const mainEmptyClass = classNames({
@@ -63,8 +65,7 @@ const Main = ({offers, currentCity, getCities, onGetCity, onGetActiveItem}) => {
                 cities={cities}
                 currentCity={currentCity}
                 // handlers
-                onGetCity={onGetCity}
-                onGetActiveItem={onGetActiveItem}
+                onGetCurrentCity={onGetCurrentCity}
               />
             }
 
@@ -107,7 +108,7 @@ const Main = ({offers, currentCity, getCities, onGetCity, onGetActiveItem}) => {
                     // properties
                     offers={offers}
                     // handlers
-                    onGetActiveItem={onGetActiveItem}
+                    onGetCurrentOffer={onGetCurrentOffer}
                   />
 
                 </div>
@@ -119,7 +120,11 @@ const Main = ({offers, currentCity, getCities, onGetCity, onGetActiveItem}) => {
 
                   {/* карта с маркерами */}
                   { offers.length > 0 &&
-                    <Map
+                    // <Map
+                    //   // properties
+                    //   offers={offers}
+                    // />
+                    <MapWrapperHoc
                       // properties
                       offers={offers}
                     />
@@ -141,7 +146,7 @@ Main.propTypes = {
   offers: PropTypes.arrayOf(PropTypes.object),
   currentCity: PropTypes.string,
   getCities: PropTypes.func,
-  onGetCity: PropTypes.func,
+  onGetCurrentCity: PropTypes.func,
 };
 
 const mapStateToProps = (state) => ({
@@ -165,13 +170,17 @@ const mapStateToProps = (state) => ({
 });
 
 const mapDispatchToProps = (dispatch) => ({
-  onGetActiveItem: (item) => {
-    dispatch(ActionCreator.getActiveItemAction(item));
-    // dispatch(ActionCreator.getCurrentOfferAction(city));
-  }
-  // onGetCity: (city) => {
-  //   dispatch(ActionCreator.changeCityAction(city));
+  // onGetActiveItem: (item) => {
+  //   dispatch(ActionCreator.getActiveItemAction(item));
+  //   // dispatch(ActionCreator.getCurrentOfferAction(city));
   // }
+
+  onGetCurrentCity: (city) => {
+    dispatch(ActionCreator.getCurrentCityAction(city));
+  },
+  onGetCurrentOffer: (offer) => {
+    dispatch(ActionCreator.getCurrentOfferAction(offer));
+  },
 });
 
 export default connect(
